Moved the models into a separate class-library
