+++ CommonLib/Models/PowershellVersion.cs
@@ -1,0 +1,8 @@
+﻿namespace CommonLib.Models
+{
+    public enum PowershellVersion
+    {
+        Legacy = 0,
+        LTS
+    }
+}
