+++ Community.PowerToys.Run.Plugin.JSLHelpers/Community.PowerToys.Run.Plugin.JSLHelpers.csproj
@@ -62,4 +62,8 @@
     <PackageReference Include="LazyCache" Version="2.4.0" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\CommonLib\CommonLib.csproj" />
+  </ItemGroup>
+
 </Project>
