+++ Community.PowerToys.Run.Plugin.JSLHelpers/Main.cs
@@ -1,7 +1,5 @@
-using System.IO;
-using System.Text.Json;
 using System.Windows.Controls;
-using Community.PowerToys.Run.Plugin.JSLHelpers.Models;
+using CommonLib.Models;
 using Community.PowerToys.Run.Plugin.JSLHelpers.QueryHandler;
 using ManagedCommon;
 using Microsoft.PowerToys.Settings.UI.Library;
