--- Community.PowerToys.Run.Plugin.JSLHelpers/PowershellVersion.cs
@@ -1,9 +1,0 @@
-﻿
-namespace Community.PowerToys.Run.Plugin.JSLHelpers
-{
-    internal enum PowershellVersion
-    {
-        Legacy = 0,
-        LTS
-    }
-}
