+++ PowerToys.sln
@@ -7,30 +7,54 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Community.PowerToys.Run.Plu
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Community.PowerToys.Run.Plugin.JSLHelpers.UnitTests", "Community.PowerToys.Run.Plugin.JSLHelpers.UnitTests\Community.PowerToys.Run.Plugin.JSLHelpers.UnitTests.csproj", "{4617AF76-31DB-429D-A794-213F1664B2CD}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CommonLib", "CommonLib\CommonLib.csproj", "{6F1B6982-D169-40F8-B10C-9CF5E06833E5}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
 		Debug|ARM64 = Debug|ARM64
 		Debug|x64 = Debug|x64
+		Release|Any CPU = Release|Any CPU
 		Release|ARM64 = Release|ARM64
 		Release|x64 = Release|x64
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{394D43B0-DD84-4590-87BF-7F650A0119F5}.Debug|Any CPU.ActiveCfg = Debug|x64
+		{394D43B0-DD84-4590-87BF-7F650A0119F5}.Debug|Any CPU.Build.0 = Debug|x64
 		{394D43B0-DD84-4590-87BF-7F650A0119F5}.Debug|ARM64.ActiveCfg = Debug|ARM64
 		{394D43B0-DD84-4590-87BF-7F650A0119F5}.Debug|ARM64.Build.0 = Debug|ARM64
 		{394D43B0-DD84-4590-87BF-7F650A0119F5}.Debug|x64.ActiveCfg = Debug|x64
 		{394D43B0-DD84-4590-87BF-7F650A0119F5}.Debug|x64.Build.0 = Debug|x64
+		{394D43B0-DD84-4590-87BF-7F650A0119F5}.Release|Any CPU.ActiveCfg = Release|x64
+		{394D43B0-DD84-4590-87BF-7F650A0119F5}.Release|Any CPU.Build.0 = Release|x64
 		{394D43B0-DD84-4590-87BF-7F650A0119F5}.Release|ARM64.ActiveCfg = Release|ARM64
 		{394D43B0-DD84-4590-87BF-7F650A0119F5}.Release|ARM64.Build.0 = Release|ARM64
 		{394D43B0-DD84-4590-87BF-7F650A0119F5}.Release|x64.ActiveCfg = Release|x64
 		{394D43B0-DD84-4590-87BF-7F650A0119F5}.Release|x64.Build.0 = Release|x64
+		{4617AF76-31DB-429D-A794-213F1664B2CD}.Debug|Any CPU.ActiveCfg = Debug|x64
+		{4617AF76-31DB-429D-A794-213F1664B2CD}.Debug|Any CPU.Build.0 = Debug|x64
 		{4617AF76-31DB-429D-A794-213F1664B2CD}.Debug|ARM64.ActiveCfg = Debug|ARM64
 		{4617AF76-31DB-429D-A794-213F1664B2CD}.Debug|ARM64.Build.0 = Debug|ARM64
 		{4617AF76-31DB-429D-A794-213F1664B2CD}.Debug|x64.ActiveCfg = Debug|x64
 		{4617AF76-31DB-429D-A794-213F1664B2CD}.Debug|x64.Build.0 = Debug|x64
+		{4617AF76-31DB-429D-A794-213F1664B2CD}.Release|Any CPU.ActiveCfg = Release|x64
+		{4617AF76-31DB-429D-A794-213F1664B2CD}.Release|Any CPU.Build.0 = Release|x64
 		{4617AF76-31DB-429D-A794-213F1664B2CD}.Release|ARM64.ActiveCfg = Release|ARM64
 		{4617AF76-31DB-429D-A794-213F1664B2CD}.Release|ARM64.Build.0 = Release|ARM64
 		{4617AF76-31DB-429D-A794-213F1664B2CD}.Release|x64.ActiveCfg = Release|x64
 		{4617AF76-31DB-429D-A794-213F1664B2CD}.Release|x64.Build.0 = Release|x64
+		{6F1B6982-D169-40F8-B10C-9CF5E06833E5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6F1B6982-D169-40F8-B10C-9CF5E06833E5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6F1B6982-D169-40F8-B10C-9CF5E06833E5}.Debug|ARM64.ActiveCfg = Debug|Any CPU
+		{6F1B6982-D169-40F8-B10C-9CF5E06833E5}.Debug|ARM64.Build.0 = Debug|Any CPU
+		{6F1B6982-D169-40F8-B10C-9CF5E06833E5}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{6F1B6982-D169-40F8-B10C-9CF5E06833E5}.Debug|x64.Build.0 = Debug|Any CPU
+		{6F1B6982-D169-40F8-B10C-9CF5E06833E5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6F1B6982-D169-40F8-B10C-9CF5E06833E5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6F1B6982-D169-40F8-B10C-9CF5E06833E5}.Release|ARM64.ActiveCfg = Release|Any CPU
+		{6F1B6982-D169-40F8-B10C-9CF5E06833E5}.Release|ARM64.Build.0 = Release|Any CPU
+		{6F1B6982-D169-40F8-B10C-9CF5E06833E5}.Release|x64.ActiveCfg = Release|Any CPU
+		{6F1B6982-D169-40F8-B10C-9CF5E06833E5}.Release|x64.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
